Clean up OptionCard
OptionCard used to have a Material 2 style shadows which is why it was
made a Card. However, we've using Material 3 with less shadows for some
time already. A decorated box fits better for visualizing the current
flat design and unlike M3 Card, it doesn't have a surprise 4px margin
around that makes it easier to align with other UI elements.